Update to run script

+++ run.rb
@@ -20,11 +20,15 @@ Dir.glob("office_2013*/*").each do |p|
   end
 end
 
+# update measures
+#cmd = "#{openstudio_path} measure -t ./measures" 
+#system(cmd)
+
 # run the osw files
 threads = []
 osw_files.each do |osw|
   if do_run
-    threads << Thread.new { cmd = "#{openstudio_path} run -w #{osw}" ; system(cmd) } 
+    threads << Thread.new { cmd = "#{openstudio_path} run --debug -w #{osw}" ; system(cmd) } 
   end
 end
 threads.each {|t| t.join}
@@ -33,20 +37,26 @@ threads.each {|t| t.join}
 results = {}
 osw_files.each do |osw|
   osw_out = File.join(File.dirname(osw), "out.osw")
-  File.open(osw_out, 'r') do |f|
-    results[osw] = JSON.parse(f.read, {symbolize_names: true})
+  if File.exists?(osw_out)
+    File.open(osw_out, 'r') do |f|
+      results[osw] = JSON.parse(f.read, {symbolize_names: true})
+    end
   end
 end
 
 # print the results
 osw_files.each do |osw|
   puts osw
-  puts "  #{results[osw][:completed_status]}"
-  results[osw][:steps].each do |step|
-    if step[:measure_dir_name] == "StandardReports"
-      step[:result][:step_values].each do |step|
-        if step[:name] == "site_energy_use_ip"
-          puts "  #{step[:value]} MBtu"
+  if results[osw].nil?
+    puts "  No results"
+  else
+    puts "  #{results[osw][:completed_status]}"
+    results[osw][:steps].each do |step|
+      if step[:measure_dir_name] == "StandardReports"
+        step[:result][:step_values].each do |step|
+          if step[:name] == "site_energy_use_ip"
+            puts "  #{step[:value]} MBtu"
+          end
         end
       end
     end
